feat: basic gosling component
